refactor(gh): extract gh/revisions submodule
Fifth and final PR of the 5-PR series splitting src/gh/mod.rs into
per-resource submodules (issue #301). Moves the gist revision-history
cluster into gh/revisions.rs: gist_commits_plan, gist_revision_plan,
fetch_gist_commits_json(_with), fetch_gist_revision_json(_with),
build_gist_revision_raw_url, fetch_revision_file_with,
fetch_revision_file_text(_with)/_optional(_with),
parse_gist_commits_json, revision_file_content, RevisionFileContent,
and their private helpers (revision_file_entry, revision_entry_raw_url,
fetch_revision_file_via_raw_url, classify_revision_file,
GhGistCommit, GhGistChangeStatus), along with their unit tests.

raw_url_fetch_plan and GhCommentUser stay in gh/mod.rs as shared
kernel (also used by gists and comments/forks respectively);
revisions.rs reaches them via
`use super::{raw_url_fetch_plan, GhCommentUser};`, matching the
convention established by the four prior submodule PRs. gh/mod.rs
re-exports revisions' public items by explicit name, so none of the
existing crate::gh::X call sites need to change.

This was the last remaining cluster in gh/mod.rs: the file is now
kernel-only (151 lines) -- parse_gh_gists, GhGist/GhGistForkOf/
GhGistFile/GhCommentUser, escape_graphql_string, raw_url_fetch_plan,
and the gh-readiness check, plus the five mod/pub-use blocks. No
signature or behavior change.

Closes #301
